mnist-flask/requirements.txt
flask==3.0.0
keras==3.10.0
tensorflow==2.18.0
pillow==10.1.0
numpy==1.26.2
gunicorn==21.2.0
